refactor(server): make use of the Plook class
